Rename demo_data_provider to data_provider
